sync/requirements.txt
requests
PyYAML
python-dotenv
tqdm
